Refactor and simplify canvas code
This commit aims to remove a lot of the duplicate drawing code,
especially for the x- and y-axes. Much of the code is unified and moved
into the root canvas class. This removes some inferior overriden
implementations. A number of unused variables and methods can also be
cleaned out.

Some work remains to be done, particularly for the top-level painting
method and the graph title.